requirements.txt
asgiref==3.10.0
Django==5.2.7
django-ckeditor==6.7.3
django-js-asset==3.1.2
djangorestframework==3.16.1
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
